Add files via upload
Added two templates that create HP and Stress trackers for Daggerheart. 
HP Stress Template.md prompts for the number of HP and Stress and creates a table with 6 rows.
HP Stress Template ITS.md prompts for the adversary name, the number of adversaries, HP, and Stress, and produces markdown that will appear as Checkboxes when the Obsidian theme ITS is installed.